requirements.txt
lxml==4.1.0
pyyaml==3.12
aiohttp==2.2.0
asyncio==3.4.3
psycopg2==2.7.3.2
PyMySQL==0.7.11
peewee==2.10.2